public/python/model_generation.py

"""
This is a sample Python implementation showing how the backend would process
image or text prompts to generate 3D models.

In a real implementation, this would be a Flask or FastAPI endpoint that receives
requests from the frontend.

Requirements:
- numpy
- trimesh
- torch
- PIL
- matplotlib
- opencv-python
"""

import os
import numpy as np
import base64
import json
from PIL import Image
import io
import trimesh
from trimesh.exchange.obj import export_obj
from trimesh.exchange.stl import export_stl
import matplotlib.pyplot as plt
from mpl_toolkits.mplot3d import Axes3D

# Simulated deep learning imports (would be real in implementation)
# import torch
# from some_package import TextTo3DModel, ImageTo3DModel

class Model3DGenerator:
    def __init__(self):
        """Initialize the 3D model generator with required models."""
        # In a real implementation, we would load the AI models here
        # self.text_model = TextTo3DModel.from_pretrained("some/model/path")
        # self.image_model = ImageTo3DModel.from_pretrained("some/model/path")
        print("Model3DGenerator initialized")
        
    def preprocess_image(self, image_path):
        """
        Preprocess the image to extract the object from background.
        
        Args:
            image_path: Path to the input image
            
        Returns:
            Processed image without background
        """
        try:
            # Load image
            img = Image.open(image_path)
            
            # In a real implementation, we would:
            # 1. Use a segmentation model to isolate the object
            # 2. Remove the background
            # 3. Normalize and resize the image
            
            print(f"Image {image_path} preprocessed successfully")
            return img
        except Exception as e:
            print(f"Error preprocessing image: {e}")
            raise
    
    def generate_from_text(self, text_prompt):
        """
        Generate a 3D model from a text prompt.
        
        Args:
            text_prompt: Text description of the object
            
        Returns:
            mesh: A trimesh object representing the 3D model
        """
        try:
            print(f"Generating 3D model from text: '{text_prompt}'")
            
            # In a real implementation, we would:
            # 1. Process the text prompt
            # 2. Feed it into the text-to-3D model
            # 3. Get back a point cloud or mesh
            
            # For demonstration, create a simple shape based on prompt keywords
            if "car" in text_prompt.lower():
                mesh = self._create_car_mesh()
            elif "chair" in text_prompt.lower():
                mesh = self._create_chair_mesh()
            else:
                # Default to a sphere
                mesh = trimesh.creation.icosphere(subdivisions=2)
            
            return mesh
        except Exception as e:
            print(f"Error generating model from text: {e}")
            raise
    
    def generate_from_image(self, image_path):
        """
        Generate a 3D model from an image.
        
        Args:
            image_path: Path to the input image
            
        Returns:
            mesh: A trimesh object representing the 3D model
        """
        try:
            # Preprocess the image to isolate the object
            processed_image = self.preprocess_image(image_path)
            
            print(f"Generating 3D model from image: {image_path}")
            
            # In a real implementation, we would:
            # 1. Process the image
            # 2. Feed it into the image-to-3D model
            # 3. Get back a point cloud or mesh
            
            # For demonstration, create a simple shape
            mesh = trimesh.creation.box(extents=[1, 1, 1])
            
            return mesh
        except Exception as e:
            print(f"Error generating model from image: {e}")
            raise
    
    def _create_car_mesh(self):
        """Create a simplified car mesh for demonstration."""
        # Body
        body = trimesh.creation.box(extents=[2, 1, 0.5])
        body.apply_translation([0, 0, 0.5])
        
        # Wheels (simplified as cylinders)
        wheel_radius = 0.2
        wheel_height = 0.1
        
        wheel_positions = [
            [0.5, 0.6, wheel_radius],
            [0.5, -0.6, wheel_radius],
            [-0.5, 0.6, wheel_radius],
            [-0.5, -0.6, wheel_radius]
        ]
        
        wheels = []
        for position in wheel_positions:
            wheel = trimesh.creation.cylinder(radius=wheel_radius, height=wheel_height)
            wheel.apply_translation(position)
            wheels.append(wheel)
        
        # Combine all parts
        car = trimesh.util.concatenate([body] + wheels)
        return car
    
    def _create_chair_mesh(self):
        """Create a simplified chair mesh for demonstration."""
        # Seat
        seat = trimesh.creation.box(extents=[1, 1, 0.1])
        seat.apply_translation([0, 0, 0.5])
        
        # Backrest
        backrest = trimesh.creation.box(extents=[1, 0.1, 1])
        backrest.apply_translation([0, -0.45, 1])
        
        # Legs
        leg_height = 0.5
        leg_width = 0.1
        
        leg_positions = [
            [0.4, 0.4, leg_height/2],
            [0.4, -0.4, leg_height/2],
            [-0.4, 0.4, leg_height/2],
            [-0.4, -0.4, leg_height/2]
        ]
        
        legs = []
        for position in leg_positions:
            leg = trimesh.creation.box(extents=[leg_width, leg_width, leg_height])
            leg.apply_translation(position)
            legs.append(leg)
        
        # Combine all parts
        chair = trimesh.util.concatenate([seat, backrest] + legs)
        return chair
    
    def save_model(self, mesh, output_path, file_format='obj'):
        """
        Save the generated 3D model to a file.
        
        Args:
            mesh: Trimesh object
            output_path: Path to save the model
            file_format: Format to save as ('obj' or 'stl')
        """
        try:
            os.makedirs(os.path.dirname(output_path), exist_ok=True)
            
            if file_format.lower() == 'obj':
                mesh.export(output_path, file_type='obj')
            elif file_format.lower() == 'stl':
                mesh.export(output_path, file_type='stl')
            else:
                raise ValueError(f"Unsupported file format: {file_format}")
                
            print(f"Model saved as {output_path}")
        except Exception as e:
            print(f"Error saving model: {e}")
            raise
    
    def visualize_model(self, mesh, save_path=None):
        """
        Visualize the 3D model using matplotlib.
        
        Args:
            mesh: Trimesh object
            save_path: Optional path to save the visualization
        """
        try:
            # Get mesh data
            vertices = mesh.vertices
            faces = mesh.faces
            
            # Create 3D plot
            fig = plt.figure(figsize=(10, 10))
            ax = fig.add_subplot(111, projection='3d')
            
            # Plot mesh
            ax.plot_trisurf(vertices[:, 0], vertices[:, 1], vertices[:, 2],
                            triangles=faces, cmap='viridis', alpha=0.7)
            
            # Set equal aspect ratio
            ax.set_box_aspect([1, 1, 1])
            
            # Remove axis labels and ticks
            ax.set_axis_off()
            
            if save_path:
                plt.savefig(save_path, bbox_inches='tight')
                print(f"Visualization saved to {save_path}")
            
            # In a web service, we wouldn't show the plot but return the image
            plt.close()
            
            return save_path
        except Exception as e:
            print(f"Error visualizing model: {e}")
            raise


def process_request(request_data):
    """
    Process a request from the frontend.
    
    Args:
        request_data: Dictionary containing request data
        
    Returns:
        Dictionary with response data
    """
    try:
        generator = Model3DGenerator()
        
        input_type = request_data.get('type')
        
        if input_type == 'text':
            text_prompt = request_data.get('data')
            mesh = generator.generate_from_text(text_prompt)
        elif input_type == 'image':
            # In a real app, the image would be uploaded and saved temporarily
            image_path = request_data.get('image_path')  
            mesh = generator.generate_from_image(image_path)
        else:
            return {'status': 'error', 'error': 'Invalid input type'}
        
        # Generate unique filenames
        import uuid
        unique_id = str(uuid.uuid4())
        
        # Save in both formats
        obj_path = f"output/{unique_id}.obj"
        stl_path = f"output/{unique_id}.stl"
        preview_path = f"output/{unique_id}_preview.png"
        
        generator.save_model(mesh, obj_path, file_format='obj')
        generator.save_model(mesh, stl_path, file_format='stl')
        generator.visualize_model(mesh, save_path=preview_path)
        
        return {
            'status': 'success',
            'obj_url': obj_path,
            'stl_url': stl_path,
            'preview_url': preview_path
        }
        
    except Exception as e:
        print(f"Error processing request: {e}")
        return {'status': 'error', 'error': str(e)}


# Example usage (as if called from a web API)
if __name__ == "__main__":
    # Example text request
    text_request = {
        'type': 'text',
        'data': 'a small toy car'
    }
    
    text_response = process_request(text_request)
    print(f"Text response: {text_response}")
    
    # Example image request (would be an actual image path in real usage)
    image_request = {
        'type': 'image',
        'image_path': 'examples/chair.jpg'  
    }
    
    image_response = process_request(image_request)
    print(f"Image response: {image_response}")
